Integrate native audio fixes with the shared Audio interface
Resubmit the remaining audio work from PR #25 on current main. Keep Sound and the shared Audio output, add native retro PCM, bounded nonblocking playback, mute controls, and exception-safe cleanup. Preserve generated ViZDoom audio logs when playback is muted.

Add English audio guidance and 24 regression tests. Refs #10.

+++ fmri_gym/adapters/base.py
@@ -42,7 +42,7 @@ class Sound:
     """
 
     pcm: np.ndarray
-    sample_rate: int
+    sample_rate: float
 
 
 class EnvAdapter:
